Add files via upload
This is a simple and user-friendly web application designed to help students and faculty members of a college to report lost items and list found items. The system acts as a centralized platform where lost belongings such as ID cards, mobile phones, books, bags, and other personal items can be easily reported and retrieved.